+++ lib/slack-ruby-bot/hooks/message.rb
@@ -3,7 +3,7 @@ module SlackRubyBot
     class Message
       def call(client, data)
         return if message_to_self_not_allowed? && message_to_self?(client, data)
-        data.text.strip! if data.text
+        data.text = data.text.strip if data.text
         result = child_command_classes.detect { |d| d.invoke(client, data) }
         result ||= built_in_command_classes.detect { |d| d.invoke(client, data) }
         result ||= SlackRubyBot::Commands::Unknown.tap { |d| d.invoke(client, data) }