+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.11.2 (Next)
 
 * [#198](https://github.com/slack-ruby/slack-ruby-bot/pull/198): Recommend using `async-websocket` instead of `celluloid-io` - [@dblock](https://github.com/dblock).
+* [#202](https://github.com/slack-ruby/slack-ruby-bot/pull/202): Allow frozen string in Hooks::Message - [@jonosenior](https://github.com/jonosenior).
 * Your contribution here.
 
 ### 0.11.1 (05/06/2018)